separate tests to improve file structure
